MassDelete.php messageManager addSuccessMessage & catch errors

+++ Controller/Adminhtml/Index/MassDelete.php
@@ -2,25 +2,36 @@
 
 namespace Cap\CleanMedia\Controller\Adminhtml\Index;
 
-use Magento\Framework\Exception\FileSystemException;
+use Exception;
 
 class MassDelete extends \Cap\CleanMedia\Controller\Adminhtml\Index
 {
     /**
      * @inheritDoc
-     * @throws FileSystemException
      */
     public function execute()
     {
-        //todo: success or failure notification
         $mediaIds = $this->getRequest()->getParams();
-        $paths = $mediaIds['selected'];
-        foreach ($paths as $path) {
-            if ($this->driverFile->isExists($path)) {
-                $this->driverFile->deleteFile($path);
+        if (!is_array($mediaIds)) {
+            $this->messageManager->addErrorMessage(__('Please select one or more medias.'));
+        } else {
+            try {
+                $paths = $mediaIds['selected'];
+                $mediaDeleted = 0;
+                foreach ($paths as $path) {
+                    if ($this->driverFile->isExists($path)) {
+                        $this->driverFile->deleteFile($path);
+                        $mediaDeleted++;
+                    }
+                }
+                $this->messageManager->addSuccessMessage(
+                    __('A total of %1 record(s) have been deleted.', $mediaDeleted)
+                );
+            } catch (Exception $e) {
+                $this->messageManager->addErrorMessage($e->getMessage());
             }
         }
 
-        return $this->_redirect('cleanmedia/*/index');
+        $this->_redirect('*/*/index');
     }
 }
